v1.1.1
Combined codes 2 (not debugged)
NOTE: I'm using a windows machine, which automatically converts LF to CRLF. As such, this program will intrinsically lose the unix-style LF after commit-checkout cycle (it will be replaced by windows-style CRLF). autocrlf is meant to handle line endings conversions, and it does, but the value needs to be manually configured, which this is not yet.